+++ Scripts/MMOGame/Database/SQLite/SQLiteDatabase_CharacterAttribute.cs
@@ -16,7 +16,7 @@ namespace MultiplayerARPG.MMO
             {
                 result = new CharacterAttribute();
                 result.dataId = reader.GetInt32("dataId");
-                result.amount = reader.GetInt16("amount");
+                result.amount = (short)reader.GetInt32("amount");
                 return true;
             }
             result = CharacterAttribute.Empty;
